requirements.txt
Pillow==9.2.0
opencv-python==4.7.0.68
imgkit==1.2.3